+++ extensions/claude/src/hooks/useModel.tsx
@@ -1,5 +1,5 @@
 import { LocalStorage, showToast, Toast } from "@raycast/api";
-import { useCallback, useEffect, useMemo, useState, useTransition } from "react";
+import { useCallback, useEffect, useMemo, useState } from "react";
 import { Model, ModelHook } from "../type";
 
 export const DEFAULT_MODEL: Model = {
@@ -34,7 +34,8 @@ async function getStoredModels(): Promise<Model[]> {
   if (!storedModels) {
     return [DEFAULT_MODEL];
   }
-  return [...JSON.parse(storedModels), DEFAULT_MODEL] satisfies Model[];
+
+  return JSON.parse(storedModels) satisfies Model[];
 }
 
 export function useModel(): ModelHook {
@@ -85,7 +86,7 @@ export function useModel(): ModelHook {
   const remove = useCallback(
     async (model: Model) => {
       const toast = await showToast({
-        title: "Remove your model...",
+        title: "Removing your model...",
         style: Toast.Style.Animated,
       });
       setData((prevData) => {
@@ -101,7 +102,7 @@ export function useModel(): ModelHook {
 
   const clear = useCallback(async () => {
     const toast = await showToast({
-      title: "Clearing your models ...",
+      title: "Clearing your models...",
       style: Toast.Style.Animated,
     });
     setData((prevData) => {